requirements.txt
transformers
datasets
# torch
peft
accelerate
bitsandbytes 
python-dotenv
sentencepiece

# DeepSpeed and its dependencies
deepspeed>=0.9.0
mpi4py>=3.1.4
ninja